+++ src/common_clj/money/converter.clj
@@ -1,0 +1,7 @@
+(ns common-clj.money.converter
+  (:require [schema.core :as s]))
+
+(s/defn ->cents :- BigInteger
+  [value]
+  (-> (* value 100)
+      bigint))