+++ CHANGELOG.md
@@ -5,6 +5,12 @@ of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
 
+## [0.4.1] - 2021-10-23
+
+### Added
+
+- Added `->cents` function were you can use to convert money amounts to cents.
+
 ## [0.3.1] - 2021-10-22
 
 ### Added
@@ -32,6 +38,8 @@ of [keepachangelog.com](http://keepachangelog.com/).
 - Add `loose-schema` function.
 
 [Unreleased]: https://github.com/macielti/common-clj/compare/0.1.1...HEAD
+
+[0.4.1]: https://github.com/macielti/common-clj/compare/0.4.1...0.3.1
 
 [0.3.1]: https://github.com/macielti/common-clj/compare/0.3.1...0.2.1